requirements.txt
asgiref                == 3.6.0
Django                 == 4.1.5
django-rest-framework  == 0.1.0
djangorestframework    == 3.14.0
pip                    == 22.3.1
pytz                   == 2022.7.1
setuptools             == 65.5.0
sqlparse               == 0.4.3
tzdata                 == 2022.7